+++ src/systems/phone/stylesheets/1024_600/dark/gesture_data.h
@@ -58,10 +58,10 @@ extern "C" {
             .direction_vertical = 50,                                               \
             .direction_horizon = 50,                                                \
             .direction_angle = 60,                                                  \
-            .top_edge = 50,                                                         \
-            .bottom_edge = 50,                                                      \
-            .left_edge = 50,                                                        \
-            .right_edge = 50,                                                       \
+            .top_edge = 10,                                                         \
+            .bottom_edge = 10,                                                      \
+            .left_edge = 10,                                                        \
+            .right_edge = 10,                                                       \
             .duration_short_ms = 500,                                               \
             .speed_slow_px_per_ms = 0.1,                                            \
         },                                                                          \